+++ packages/orq-rc/src/orq_ai_sdk/models/updatecontactop.py
@@ -181,7 +181,7 @@ class UpdateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-15T12:57:20.627Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-15T16:01:51.772Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")